inacanoe/qa/web_src/requirements.txt
flask
redis
psycopg2-binary
sqlalchemy
Flask-SQLAlchemy
